+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#3784](https://github.com/clojure-emacs/cider/issues/3784): Inspector: make point less erratic when navigating between inspector screens.
 - [#3790](https://github.com/clojure-emacs/cider/issues/3790): Stacktrace: show messages and data for all exception causes by default.
 - [#3789](https://github.com/clojure-emacs/cider/issues/3789): Refactor and simplify exception handling.
+- [#3359](https://github.com/clojure-emacs/cider/pull/3359): Add customizable default connection params
 
 ## 1.17.1 (2025-02-25)